organize tag components
