+++ lib/content-services/src/lib/tag/tag-actions/tag-actions.component.ts
@@ -17,7 +17,7 @@
 
 import { TranslationService } from '@alfresco/adf-core';
 import { Component, EventEmitter, Input, OnChanges, Output, ViewEncapsulation, OnDestroy, OnInit } from '@angular/core';
-import { TagService } from './services/tag.service';
+import { TagService } from '../services/tag.service';
 import { Subject } from 'rxjs';
 import { TagPaging } from '@alfresco/js-api';
 import { takeUntil } from 'rxjs/operators';
