+++ lib/content-services/src/lib/tag/tag-list/tag-list.component.spec.ts
@@ -16,10 +16,10 @@
  */
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TagService } from './services/tag.service';
-import { TagListComponent } from '././tag-list.component';
+import { TagService } from '../services/tag.service';
+import { TagListComponent } from './tag-list.component';
 import { of } from 'rxjs';
-import { ContentTestingModule } from '../testing/content.testing.module';
+import { ContentTestingModule } from '../../testing/content.testing.module';
 
 describe('TagList', () => {
     const dataTag = {
